requirements.txt
aiohttp==3.5.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
future==0.17.1
gspread==3.1.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.5.2
oauth2client==4.1.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
yarl==1.3.0
